move internal methods to soft_keyboard_shortcut_support_internal, remove try/catch
